backend/requirements.txt
django==3.0.7
djangorestframework==3.11.0
django-modeltranslation==0.15.0